Convert Graph back to storing edge weights as floats, rather than doubles - precision wasn't the problem, bidirectionality was

+++ src/main/scala/software/uncharted/graphing/clustering/unithread/Community.scala
@@ -436,11 +436,11 @@ class Community (val g: Graph,
     val degrees = new Array[Int](nb_nodes)
     val links = MutableBuffer[Int]()
     val nodeInfos = new Array[NodeInfo](nb_nodes)
-    val weights = MutableBuffer[Double]()
+    val weights = MutableBuffer[Float]()
 
     val comm_deg = comm_nodes.size
     for (comm <- 0 until comm_deg) {
-      val m = MutableMap[Int, Double]()
+      val m = MutableMap[Int, Float]()
 
       val comm_size = comm_nodes(comm).size
       for (node <- 0 until comm_size) {
@@ -449,7 +449,7 @@ class Community (val g: Graph,
         for (i <- 0 until deg) {
           val (neigh, neigh_weight) = neighbors.next
           val neigh_comm = renumber(n2c(neigh))
-          m(neigh_comm) = m.getOrElse(neigh_comm, 0.0) + neigh_weight
+          m(neigh_comm) = m.getOrElse(neigh_comm, 0.0f) + neigh_weight
         }
       }
       degrees(comm) = if (0 == comm) m.size else degrees(comm-1) + m.size
