+++ src/main/scala/software/uncharted/graphing/clustering/unithread/GraphEdges.scala
@@ -198,9 +198,9 @@ class GraphEdges (val links: Array[_ <: Seq[(Int, Float, Seq[String])]]) {
 
     // Allocate destination and weight arrays
     val edgeDestinations: Array[Int]           = new Array(totalEdges)
-    val edgeWeights:      Option[Array[Double]] =
+    val edgeWeights:      Option[Array[Float]] =
       if (weighted) {
-        Some(new Array[Double](totalEdges))
+        Some(new Array[Float](totalEdges))
       } else {
         None
       }
